append 继续完善 request微服务

+++ commons/src/main/java/top/camsyn/store/commons/entity/user/User.java
@@ -31,7 +31,7 @@ public class User extends Model<User> {
 
     private Integer credit;
 
-    private Integer liyuan;
+    private Double liyuan;
 
     private String headImage;
